+++ src/BaseAsset.php
@@ -2,8 +2,6 @@
 
 namespace Inpsyde\Assets;
 
-use Inpsyde\Assets\Exception\InvalidArgumentException;
-
 abstract class BaseAsset implements Asset
 {