+++ tests/fixtures/asset-config.php
@@ -1,16 +1,20 @@
 <?php
 
 use Inpsyde\Assets\Asset;
+use Inpsyde\Assets\Script;
+use Inpsyde\Assets\Style;
 
 return [
     [
         'handle' => 'foo',
         'url' => 'foo.css',
-        'type' => Asset::TYPE_STYLE,
+        'type' => Asset::FRONTEND,
+        'class' => Style::class
     ],
     [
         'handle' => 'bar',
         'url' => 'bar.js',
-        'type' => Asset::TYPE_SCRIPT,
+        'type' => Asset::FRONTEND,
+        'class' => Script::class
     ],
 ];